清除所有数据接口仅在开发环境暴露&升级scirichon-token-checker增加alias字段&用例维护(1h)

+++ routes/index.js
@@ -7,11 +7,13 @@ const discussions = require('./discussion')
 const router = new Router();
 router.use(`${base_route}/articles`,  articles.routes(),articles.allowedMethods())
 router.use(`${base_route}/discussions`,  discussions.routes(),discussions.allowedMethods())
-router.del(`${base_route}/synergy`,async function(ctx) {
-    await(dbHelper.pool.query(`delete from "Articles"`));
-    await(dbHelper.pool.query(`delete from "Discussions"`));
-    await(dbHelper.pool.query(`delete from "ArticleScores"`));
-    ctx.body = {}
-})
+if(process.env.NODE_ENV === 'development') {
+    router.del(`${base_route}/hidden`, async function (ctx) {
+        await(dbHelper.pool.query(`delete from "Articles"`));
+        await(dbHelper.pool.query(`delete from "Discussions"`));
+        await(dbHelper.pool.query(`delete from "ArticleScores"`));
+        ctx.body = {}
+    })
+}
 
 module.exports = router
